requirements.txt
asgiref==3.6.0
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.2
decorator==5.1.1
Django==4.1.7
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
dynaconf==3.1.12
flake8==6.0.0
hvac==1.1.0
idna==3.4
loguru==0.6.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pyhcl==0.4.4
pytz==2022.7.1
requests==2.28.2
shortuuid==1.0.11
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.15
validators==0.20.0
